+++ Sources/PitwallApp/AppDelegate.swift
@@ -1,45 +1,17 @@
 import AppKit
-import PitwallAppSupport
 
 final class AppDelegate: NSObject, NSApplicationDelegate {
-    private var statusItem: NSStatusItem?
+    private var menuBarController: MenuBarController?
 
     func applicationDidFinishLaunching(_ notification: Notification) {
         NSApp.setActivationPolicy(.accessory)
-        configureStatusItem()
+        let controller = MenuBarController()
+        controller.start()
+        menuBarController = controller
     }
 
     func applicationWillTerminate(_ notification: Notification) {
-        statusItem = nil
-    }
-
-    private func configureStatusItem() {
-        let item = NSStatusBar.system.statusItem(withLength: NSStatusItem.variableLength)
-        item.button?.title = "Pitwall"
-        item.button?.toolTip = "Pitwall provider pacing"
-        item.menu = makeMenu()
-        statusItem = item
-    }
-
-    private func makeMenu() -> NSMenu {
-        let menu = NSMenu()
-
-        let titleItem = NSMenuItem(
-            title: "Pitwall",
-            action: nil,
-            keyEquivalent: ""
-        )
-        titleItem.isEnabled = false
-        menu.addItem(titleItem)
-        menu.addItem(.separator())
-        menu.addItem(
-            NSMenuItem(
-                title: "Quit Pitwall",
-                action: #selector(NSApplication.terminate(_:)),
-                keyEquivalent: "q"
-            )
-        )
-
-        return menu
+        menuBarController?.stop()
+        menuBarController = nil
     }
 }